Moved Footer into Settings file
