+++ Jewel/Views/Options/Settings.swift
@@ -58,3 +58,16 @@ struct Settings: View {
     .navigationViewStyle(StackNavigationViewStyle())
   }
 }
+
+struct Footer: View {
+  var body: some View {
+    VStack {
+      Text("🎵 + 📱 = 🙌")
+        .padding(.bottom)
+      Text("© 2020 Breakbeat Ltd.")
+      Text(Bundle.main.buildNumber)
+        .foregroundColor(Color.secondary)
+    }
+    .font(.footnote)
+  }
+}
